requirements.txt
isort==5.11.4
numpy==1.24.1
pandas==1.5.2
passlib==1.7.4
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
